I pared the library down by literally 1.4k lines of code. Yeah, this is good work. I almost feel like starting work on the optimized plastic RNN, but I've been awake for almost 10h by now. This is no time to be starting new things. Let me rest for today and tomorrow I will go at it full blast. I'll aim to deal with the matrix optimizations and PRONG.
This will be challenging to do and I am not sure I will be able to manage it all tomorrow, but once I do I will have mastered the material in: https://devblogs.nvidia.com/optimizing-recurrent-neural-networks-cudnn-5/

Being able to do this elegantly was one of the major motivations in me creating Spiral, in fact I think this is the last thing on my TODO list from 2016 that I haven't yet done. I've already done the multiplicative integration, layer norm and the wavefront iteration. Now only this is left.

I will go at it as soon I recharge.