https://arxiv.org/pdf/1502.03492.pdf Gradient-based Hyperparameter Optimization through Reversible Learning
I've been thinking about this paper for the past 1.5 hour. Miconi linked it to me. I guess I'll let go of my grudge for not telling me about reversible learning earlier.

During the night I realized that it is not just Oja's rule. There is a large number of potential learning rules that could benefit greatly from reversible learning.

For example In the EGHR paper, I could make the ICA gradient term the prediction error for the next output like in the char RNN. There might be ways of doing online whitening in a recurrent net too. Some of the Hebbian learning rules that I've seen do that.

It is like a lightbulb went on in my head. Three years ago when I first saw this paper I could not see a use for it, but reversible learning is in fact the key to everything.

Unfortunately, as it often turns out in time like these it seems that doing it like I have imagined will cause the errors to accumulate exponentially. I believe the paper in this regard.

I also need a better solution to this than the one they propose in the paper.

I said yesterday I was going to do the optimized implementation of the plastic_rnn today, but let me take some time off from that to do research.

PRONG and Hebbian learning are the source. They are the two biggest things I need to master in 2018.

If I can grasp the second one properly, I will finally have something that will put me beyond the curve.

https://pdfs.semanticscholar.org/2a0e/6578f71793d29f8272afae75f9d610e1da46.pdf
An Investigation of the Gradient Descent Process in Neural Networks

Let me start with Pearlmutter's thesis and then I will look into reversible computation. It is not hugely important at the moment, but it will be #1 priority in the future.

I want to see what this thesis says about errors. The paper I linked to earlier referenced it.