I did some experimentation in order to study accumulation of errors over multiple timesteps. The inputs and outputs are drawn from a uniform distribution. Here is the F# code that I used to print this out.
```
type R<'a> = {passes: int32; error : 'a}

let inline error_template float passes H' H =
    let rng = System.Random()
    let input_out = Array.init passes (fun _ -> rng.NextDouble() |> float, rng.NextDouble() |> float)

    let Hs, r = Array.mapFold H' (float 0.0) input_out
    let H's, r' = Array.mapFoldBack H input_out r

    let errors = Array.map2 (fun a b -> abs (a - b)) Hs H's
    {passes=passes; error=errors.[0]} // The error is highest at the beginning as could be expected.

let inline error_oja float passes =
    let n = float 0.01
    let one = float 1.0

    let H' H (input, out) =
        let r = H + n * (input * out - out * out * H)
        r,r
    let H (input, out) H' =
        let r = (n * input * out - H') / (n * out * out - one)
        H',r

    error_template float passes H' H

let inline error_ema float passes =
    let n = float 0.01
    let one = float 1.0

    let H' H (input, out) =
        let r = n * input * out + (n - one) * H
        r,r
    let H (input, out) H' =
        let r = (H' - n * input * out) / (n - one)
        H',r

    error_template float passes H' H

let passes = [|500;1000;2000;3000;4000;5000;6000;7000;8000;9000|]
printfn "The Oja errors with float32 are:"
Array.iter (error_oja float32 >> printfn "%A") passes

printfn "The Oja errors with float64 are:"
Array.iter (error_oja float >> printfn "%A") passes

printfn "The EMA errors with float32 are:"
Array.iter (error_ema float32 >> printfn "%A") passes

printfn "The EMA errors with float64 are:"
Array.iter (error_ema float >> printfn "%A") passes
```

Here is what this prints out.

```
The Oja errors with float32 are:
{passes = 500;
 error = 2.57045031e-07f;}
{passes = 1000;
 error = 2.0879088e-07f;}
{passes = 2000;
 error = 4.14921669e-05f;}
{passes = 3000;
 error = 0.000161794596f;}
{passes = 4000;
 error = 0.0458912589f;}
{passes = 5000;
 error = 1.71386075f;}
{passes = 6000;
 error = 39.9035683f;}
{passes = 7000;
 error = 1335.82397f;}
{passes = 8000;
 error = 6237.02881f;}
{passes = 9000;
 error = 157804.75f;}
The Oja errors with float64 are:
{passes = 500;
 error = 9.207044849e-16;}
{passes = 1000;
 error = 2.262426357e-14;}
{passes = 2000;
 error = 3.600440258e-13;}
{passes = 3000;
 error = 2.779200134e-11;}
{passes = 4000;
 error = 2.532183881e-11;}
{passes = 5000;
 error = 1.356762178e-08;}
{passes = 6000;
 error = 8.418838118e-09;}
{passes = 7000;
 error = 9.325236929e-06;}
{passes = 8000;
 error = 0.0001536395682;}
{passes = 9000;
 error = 0.001170499381;}
The EMA errors with float32 are:
{passes = 500;
 error = 1.43889338e-07f;}
{passes = 1000;
 error = 3.03070992e-05f;}
{passes = 2000;
 error = 0.975888908f;}
{passes = 3000;
 error = 19528.3086f;}
{passes = 4000;
 error = 297029792.0f;}
{passes = 5000;
 error = 7.44508411e+12f;}
{passes = 6000;
 error = 4.23176737e+17f;}
{passes = 7000;
 error = 3.88952453e+21f;}
{passes = 8000;
 error = 6.5476209e+26f;}
{passes = 9000;
 error = 2.97080645e+30f;}
The EMA errors with float64 are:
{passes = 500;
 error = 2.545706701e-16;}
{passes = 1000;
 error = 4.65724681e-13;}
{passes = 2000;
 error = 2.369793283e-09;}
{passes = 3000;
 error = 3.540763222e-05;}
{passes = 4000;
 error = 0.359336145;}
{passes = 5000;
 error = 13473.68878;}
{passes = 6000;
 error = 441041068.4;}
{passes = 7000;
 error = 4.057630071e+12;}
{passes = 8000;
 error = 2.700332248e+16;}
{passes = 9000;
 error = 1.864570819e+21;}
```

This is just one sample. The way the errors accumulate is not really linear. Sometimes the errors do not go down much and sometimes there are huge increases in error, but it is never worse than multiplying by e2 for the Oja's rule.

The paper has got me scared about numerical inaccuracies so I've made a little proxy to measure what I can expect for a single parameter in terms of error. It is important to keep in mind that a single layer Hebbian matrix whose updates only depend on its input and output, and its weight parameter will be much more stable than what they do in that paper. In that paper they have to account for reversing the learning for the whole network which is a much more difficult task than having to reverse learning in isolated layers. In addition to that, regardless what they use as the network, the hyperparameters they optimize will generally be scalars meaning the errors will become concetrated in them. This is a much different situation than in the whole matrix where the errors will be distributed.

Furthemore some forms of Hebbian learning might be more stable than others. Oja definitely seems more stable than EMA here.

Using float64 numbers over float32 ones in the plastic matrix definitely seems to be a worthwhile tradeoff, but up to 1500 timesteps I'd say it would be safe to use even float32. For float64 I'd say 7k would be the limit.

I think the toy experiment demonstrates that reversible learning might be a viable method for taking advantage of Hebbian learning in NNs.

The stuff above is part of the email I sent to Miconi. Let me take a break here. This is actually quite encouraging. I do not think reversible learning will be a problem in practice.

I still want to study this more to see if I can make it more stable somehow. I still haven't even touched Pearlmutter's thesis and the unsupervised NLP paper by OpenAI.