I am done with this. The errors with weight norm accumulate linearly and they are small even up to 200k timesteps.
All in all, I've done well to discover this. In one day I've solved a major issue scaling issue with plasticity and in the next I've solved a major issue to that solution. Now everything is solved and I can be happy about my future prospects. Differentiable plasticity will surely be my friend.