"9:35am. https://igi-web.tugraz.at/people/maass/publications.html
Oh, he has some Berkley lectures.

https://www.youtube.com/watch?v=zpDzSonhymE&index=7&list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A

Maybe this was related to the course. At any rate, I feel like watching videos today. Let me skim the thesis and then I will watch this. Or actually, maybe I'll watch the video on reversible computation first.

9:55am. https://www.coursera.org/learn/synapses#syllabus

Ah, I think this was the course I was talking about.

Done with the thesis. Forget it. I do not feel like drilling into it.

10:05am. https://youtu.be/zpDzSonhymE?list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A&t=588

"In addition hardly any two neuron in the brain have the same input/output behavior."

10:15am. https://youtu.be/zpDzSonhymE?list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A&t=1141

This stuff is interesting. It is also interesting that constraint satisfaction is popping up again. I saw it in Pearlmutter's thesis. But I agree that the brain would definitely need to do this.

Let me resume.

10:45am. https://youtu.be/zpDzSonhymE?list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A&t=2431

Hmmm, is there a paper where these experiments were done. I'd be interested in seeing it.

11am. https://youtu.be/zpDzSonhymE?list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A&t=3221

He gives a bunch of references here. He also has another talk on this subject, but I want to watch the reversible computation one. Let me do it for 20m and then I will go get breakfast.

https://www.youtube.com/watch?v=IQZ_bQbxSXk
Stanford Seminar - Generalized Reversible Computing and the Unconventional Computing Landscape

It is this Stanford seminar that I had in mind."