Reversible computing flipped some switches in my brain it seems. Well to be fair, I am not sure whether it would be possible to do this kind of reversibility - I am not an expert on this. But I know what I should aim for
Now that I am done with the reversible computing video what should I do next? I guess, I should go for the second video by Wolfgang Mass. Let me go for it and then I might call it a day.

The following is my second reply to /u/gwern.

"> What kind of hardware for running what, exactly?

Let me try to tackle this question again as I've had a vision of the future that seems reasonable, but is off the beaten path.

https://www.youtube.com/watch?v=IQZ_bQbxSXk
Stanford Seminar - Generalized Reversible Computing and the Unconventional Computing Landscape

I've known about reversible computed for a while now and this video lays out the case nicely that is the only way forward that tackles the limits of adiabatic computing. He is right of course - I'd like nonvolatile memories and think they will certainly arrive before reversible computing, but physics are physics.

The reason I am trying this question again is because I've just realized that in the context of machine learning there are significant metalearning applications that would be uncovered with the ability to use [reversible learning](https://arxiv.org/abs/1502.03492).

In that paper they use it to optimize hyperparameters which is a dull application of reversible learning. What could be done instead is to use it to reverse the weight updates of subnetworks like in [differentiable plasticity](https://arxiv.org/abs/1804.02464) rather than make copies of them at every timestep.

While working on differentiable plasticity I've run into the issue of them being huge memory guzzlers as they need to make copies of the their plastic weights on each timestep in order to backprop through them. This severely limits their real world applicability, but with reversible learning it would be possible to get around this fatal failure completely. It would also completely obliviate the need for those numerical stability hacks in the AD paper which is the real limitation of reversible learning at the moment. Ditching floating point arithmetic for fixed point and using information buffers like they propose is just not in any way realistic.

In simple unsupervised architectures like the one from the DP paper it is actually possible to use reversible learning even today by combining Hebbian learning with weight normalization. Experimentally, I've verified that the Oja and the EMA rules are unstable - they can only go up to 1k-10k timesteps depending on whether EMA or Oja's rule is used (Oja is more stable) and whether float32 or float64 types are used. After that the errors accumulate exponentially. This still is not a small number of timesteps, but it turns out that Hebbian learning updates with weight normalization can be reversed for 100s of thousands of timesteps with only slight, barely linear accumulation of error.

I've also verified that adding modulation to that architecture makes it work over an order of magnitude better. Though that is only on the Binary Pattern test from the paper - should that generalize then RNNs with Hebbian learning should become by far the most dominant RNNs for the nearby future given that already 1 layer unmodulated plastic RNNs are competitive with huge architectures like SNAIL on some tasks.

But actually the concept of reversible learning is actually bigger than that - even beyond the allowing far wider range of learning rules and architectures to be safely used, wouldn't being able to reverse computation allow backpropagation to be used for arbitrary number of timesteps without requiring the storage of intermediaries apart from inputs? That would be quite something.

Given its noisiness, this is actually a solution that the brain cannot possibly be using even though it is estimated to be capable of storing a lifetime worth memories in high fidelity. It would be a solidly superhuman capability."