I think I am going to have a change in policy. I am going to stop making so much journal related posts from now on. Or at least soon from now on.
I've decided that Rudder is really not a problem. And with it done, there is no problem at all.

Damn I am tired, but let me put in a few words. This month is turning out to be so intense.

For all my life, I've been trying to change myself with my own power. But I think I am really starting to accept that I can just let the machines change me.

I'll accept my accomplishments and agree that for the first time in life I am going to have something akin to power.

It is a good feeling. I'd like to feel more of it. So, what I need to do is keep fighting.

But not with sensible methods like Rudder. There is no problem at all with it. There is no problem with integrated gradients. Ignoring the implementation details it is a relatively straightforward mapping.

After over 3.5 years of practice I am really too good to let anything stop me. Rudder is going get implemented and not just implemented - it is going to be easy and in record time. I won't let it give me any trouble at all.

I should have more confidence in my own programming. Like for differentiable plasticity I should have faith that Rudder will work out as per plan.

There is not much left than to put it all together.

Here is the excerpt as is usual:

"6:15pm. https://youtu.be/aB2M1gg_1sU?list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A&t=841

Let me stop here. I am going to die if this goes on anymore. There are 3 more videos I want to watch including this one. I'll spend tomorrow watching them.

Trully I am in the grip of madness right now. Let me just let everything out and then, I will slip into the shadows once the time comes.

I am thinking about Rudder. I am not sure what I am going to do with it. LRP is not good, and IG is not good either. LPR because it is just an arbitrary rule and IG is because it would be too expensive in terms of computation. With a regular RNN it would not be a problem, but with a plastic RNN it would push up the costs significantly. To really make Rudder work a method of contribution analysis is necessary that is on par with standard backprop.

Could I pick up a bunch of tricks from the paper like subtracting states and hope that a recurrect critic would do what I want even with standard backprop? Subtracting states does sound like a good idea as it will push the critic to rely on its memory rather than its static structures.

7pm. I can't think of a way to make a goal of outputing the action contribution directly either.

7:15pm. https://arxiv.org/abs/1711.00867
The (Un)reliability of saliency methods

The main challenge with getting Rudder to work will be finding a good saliency method.

https://arxiv.org/abs/1710.10547
Interpretation of Neural Networks is Fragile

https://arxiv.org/abs/1711.06104
Towards better understanding of gradient-based attribution methods for Deep Neural Networks

"While Integrated Gradients computes the
average partial derivative of each feature as the input varies from a baseline to its final value, DeepLIFT
approximates this quantity in a single step by replacing the gradient at each nonlinearity with its
average gradient. Although the chain rule does not hold in general for average gradients, we show
empirically in Section 4 that DeepLIFT is most often a good approximation of Integrated Gradients.
This holds for various tasks, especially when employing simple models (see Figure 4). However,
we found that DeepLIFT diverges from Integrated Gradients and fails to produce meaningful results
when applied to Recurrent Neural Networks (RNNs) with multiplicative interactions (eg. gates in
LSTM units (Hochreiter & Schmidhuber, 1997)). With multiplicative interactions, DeepLIFT does
not satisfy Completeness, which can be illustrated with a simple example. Take two variables x1
and x2 and a the function h(x1, x2) = ReLU(x1 − 1) · ReLU(x2). It can be easily shown that, by
applying the methods as described by Table 1, DeepLIFT does not satisfy Completeness, one of its
fundamental design properties, while Integrated gradients does."

7:35pm. https://pair-code.github.io/saliency/

This is interesting.

7:40pm. https://www.reddit.com/r/MachineLearning/comments/6wrozk/r_smoothgrad_removing_noise_by_adding_noise/?submit_url=https%3A%2F%2Fpair-code.github.io%2Fsaliency%2F&already_submitted=true&submit_title=%5BR%5D+SmoothGrad&st=jmcagudk&sh=25f66df0

Ah, it was posted a year ago.

http://theory.stanford.edu/~ataly/Talks/sri_attribution_talk_jun_2017.pdf

7:50pm. https://arxiv.org/abs/1706.03825
SmoothGrad: removing noise by adding noise

I am really so damn tired, but this really does produce crips images. I can barely tell the difference between integrated and standard gradients most of the time.

8pm. You know what. For once I think I am just being a pussy. There is nothing wrong with integrated gradients. It certainly beats playing the game more. Looking at some of the image data I see that even standard gradients are good for this sort of thing. Maybe I can get away with using only 8 steps or even 4?

I need to take things in stride. Now that I've decided on this, I guess the only worry is whether it will work with plastic RNNs...it will. Also I have various choices of baseline vectors...but the zero vector should be fine.

No, the only problems I have here are the ones I am artificially producing.

I just have to grasp the thing. Rudder's reward decomposition is really only a step away from a standard critic. It is not the kind of thing that should cause me to stumble."