Rudder will depend completely on whether I can take only a few IG steps, and whether I won't need to do hacks like setting the rewards in the last timesteps to zero.
I will give it a shot and ditch it if it does not work out. In that case I will simply wait for a better contribution method than IG to appear. That will be it.

Let me make a commit and then I will start. I want to go through the neuroscience videos today, but I might continue this into tomorrow if I decide to start the course.