I had enough. I'll watch the rest of the course tomorrow.
https://arxiv.org/abs/1805.08394
State-Denoised Recurrent Neural Networks

https://arxiv.org/abs/1802.10542
Memory-based Parameter Adaptation

Let me past the two papers here. The first one seems interesting, but I do not have the energy to study it now.

I'll save it for tomorrow. Uah, how annoying. It seems that when poker is said and done, I will be working on adding higher order differentiation to the library in order to support reversible learning. That will be the only path in the end.

I wonder where that leaves Hebbian learning? Where does that leave reward decomposition?

How far will I have to go to reach the Singularity? Well, I do not expect it to be able to reach it on my own. It will require me to rely on the community to fill in the gaps as they occur.

...Ah, I am not sure to be honest. Given how powerful RNNs with Hebbian learning are expected to be, would there really by anything to gain by going down that avenue? I cannot reason it through. Hebbian learning needs to be somewhere because it is the most sensible form of unsupervised learning to use and has powerful connections to biology, but how far the hole goes is a mystery to me.

I am still worried about Rudder. What I wrote yesterday about me needing to just implement it does not stick. I do not entirely trust IG, especially when combined with memory and RNNs. I mean, is passing in the wrong inputs and summing the gradients really the right way to calculate contributions?

Even if it works with static RNNs, just what will happen once I bring in plastic ones? I just do not believe it.

First of all, what I am going to impose is that the state gets passed in as is. I really should be looking at the difference in gradients when the action is given and when it is not and integrate over that.

Furthermore, to solve a recurring problem that I seem to be having, I am going to be introducing special functions that replicate input. By recurring problem, I am alluding that I am having this same problem with the biases for PRONG and I haven't solved it adequately. Rather than make the IG special I think I am going to just make it a special node.

...I think that should work to simplify things.

...But no it won't be enough. Ugh, this is beyond annoying. With RNNs if I am going to be using integrated gradients, I really should avoid calculating the adjoints for the weights. I only really need them for the inputs. Yeah, it is not going to work - this is going to require some magic. But it will be doable. Regular ACs require special considerations too. Rudder is just a upgraded critic.

I originally put the learning rates for PRONG during the backward pass in part because I wanted a way to disable it when doing Rudder, but I am going to have to go a bit further than that it seems. I originally intended to use the PRONG backward pass trick for the plastic matrix and spent a lot of time thinking about it, but then it turned out to be a complete bust. And now I am going to use the K-FAC instead of PRONG as it would be cheaper ironically. Well, probably it would not be as expensive as I'd feared, all those traversals over the weight matrix are what is truly expensive. If for steps of only a few dozen, K-FAC should be on par with PRONG.