I feel like making a huge, but simple gamble to grasp the true hope that lies beyond the despair. If it works it will just save me so much effort it is not even funny. It would be like winning the lotery in a sense, it will make it 100% certain that I will fullfil all my goal for the year and pave the way to grasping the Singularity.
If it does not work, I will have to go back to thinking about Rudder, but if it works then it will push me to the head of the pack in heartbeat and field of RL will have to be rewritten to accomodate the new reality of how credit assignment should be done. The old algorithms that they've been working up to now will go the way of the horses.

I will be able to unleash the full power of RNNs at last.

"1:05pm. I am thinking that given the plastic RNN's shocking behavior with regard to the output, it might be the associative operation that I've been seeking to make the action/state chain that I can optimize without resorting to MC or TD. The cats cannot learn to see without action is just another clue that the unified view that I've been seeking months ago all the way when I started might be the correct direction to go in.

1:25pm. I am a bit too distracted right now for the lectures. Maybe I will have to continue them tomorrow if I fail to finish them.

I should seriously consider that the only solution would be too use some kind of plastic RNN as the action layer and use that to link the action chain. Commonsense reasoning like I've been doing months ago will get me nowhere. All I can do in this situation is open my mind to the fullest and let the creative mania take hold. It will lead me to the truth in a way nothing else can.

I know that standard RNNs would be useless for this, but looking at what the modulated RNN is doing on the Binary Test is really making it obvious that it is different. It would be worth testing this hypothesis at least.

2pm. I am thinking some more. Even if the above worked, unlike with MC, I would have no way of making the chain integrous. That is, if one of the actions become 100%, then no reward would be propagated even if the rest of the path is not solid. Maybe that does not really matter. As long as the tendency is there that might be enough. Also...maybe I could use something other than a softmax?

2:05pm. But for plastic RNNs, it is not a given that they won't be integrous. The thing Wolfgang said about noisiness of the brain is coming back to me. I think that as long as there is enough stochasticity, there will be integrousness. And complete integrousness is not the point as it would reduce exploration.

2:10pm. I think that integrousness could also be called approximate optimality. Levine talked about that in the Bayesian part of his course. Like, full out MC or TD would have a tendency to converge to a deterministic policy if one of the outcomes is optimal even if the intermediate actions make no difference. This absolutely is not biologically plausible. It would happen to all the current RL algorithms, but not with what I am proposing.

2:15pm. Actually, if anything this flaw of my proposal actually makes it much more likely that it is right. Wolfgnang said that the brain might produce noise, but that cannot be the case - it is more likely that is plasticity leads to chaotic behavior. That would be ideal for exploration as a side effect which is related to integrousness.

2:45pm. The more I think about it, the more I come to the conclusion that with Hebbian learning, this might be doable. It is gamble, but I haven't seen anything to prove yet that what I want to do is impossible.

I think that even with Rudder addressing the issue of delayed rewards the noise of MC and TD that would be needed to be done on top of it would be overwhelming for long time horizons."

My feelings are starting to turn towards the ideas I had back in April and May, and now that I have this new tool I feel I should try them again.

...No actually, there are two tools that I have. The first would be the new RNN. But the second would be the state difference trick from the Rudder paper.

I cannot forget about it. It might turn out to be crucial wedge to pry the door open for the new credit assignment scheme. Unless I use it, the network might be too tempted to use its static pathways to do learning and I want to prevent that. It might be the difference between victory and defeat. I absolutely must not forget it.

...And I must believe in this madness that is guiding me.

Phew. With that settled, I will focus on this tomorrow.

Somehow today I was so much in thought that I barely got much studying in at all. I feel like those studies are just excuses so I can turn off from being goal directed and let my mind spend my time cultivating ideas.

I thought a fair bit about Rudder as well, but if I can avoid it for the idea I just mentioned that would be great.

The method has a fair bit of disadvantages steming for having to use integrated gradients to do the contribution analysis. I also thing that the way they do it in the paper by predicting the end reward is wholy unnecessary. Redistributing the individual reward would be much lower variance.

If I only could make the experiment a success it will clear such a large amount of burden...

Ultimately, I should just do what I want. Rather than hoping I should just decide that I want the RNN to do what I want and go forward with the expectation what reality will have to match up with what I want it to be.

That would be the most what a true seeker would be like. I can just as easily turn around.

There is no need to be reasonable. Even though there is no real reason to think that what I want will succeed - the effect of these algorithms is devilishly hard to predict, nonetheless I do not feel at like I am doing the lotery.

Rather it feel like 100% that this should be the case that this will work. It feels like if this is wrong then nothing can possibly be right.

Ultimately, progress in ML does not have to take a linear path - new and better models should be different at a qualitative level. They have every reason to enable new ways of usage and fundamentally new and better capabilities that the old tools cannot even be imagined to have.