"9:40am.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 248 and 752 out of 1000.
The time was 00:00:07.9802642 for: iteration 0
Starting timing for: iteration 1
Winrate is 269 and 731 out of 1000.
The time was 00:00:10.4196579 for: iteration 1
Starting timing for: iteration 2
Winrate is 397 and 603 out of 1000.
The time was 00:00:13.5186576 for: iteration 2
Starting timing for: iteration 3
Winrate is 526 and 474 out of 1000.
The time was 00:00:14.3051515 for: iteration 3
Starting timing for: iteration 4
Winrate is 625 and 375 out of 1000.
The time was 00:00:15.3061591 for: iteration 4
Starting timing for: iteration 5
Winrate is 650 and 350 out of 1000.
The time was 00:00:14.2120740 for: iteration 5
Starting timing for: iteration 6
Winrate is 660 and 340 out of 1000.
The time was 00:00:14.2610872 for: iteration 6
Starting timing for: iteration 7
Winrate is 699 and 301 out of 1000.
The time was 00:00:13.9698238 for: iteration 7
Starting timing for: iteration 8
Winrate is 693 and 307 out of 1000.
The time was 00:00:14.3638408 for: iteration 8
Starting timing for: iteration 9
Winrate is 678 and 322 out of 1000.
The time was 00:00:14.5854921 for: iteration 9
Starting timing for: iteration 10
Winrate is 717 and 283 out of 1000.
The time was 00:00:13.6562069 for: iteration 10
Starting timing for: iteration 11
Winrate is 731 and 269 out of 1000.
The time was 00:00:13.8396525 for: iteration 11
Starting timing for: iteration 12
Winrate is 723 and 277 out of 1000.
The time was 00:00:13.6616550 for: iteration 12
Starting timing for: iteration 13
Winrate is 720 and 280 out of 1000.
The time was 00:00:13.6705641 for: iteration 13
Starting timing for: iteration 14
Winrate is 702 and 298 out of 1000.
The time was 00:00:14.6458001 for: iteration 14
```

Well, if a feedforward net can do it I do not see why the plastic RNN could not.

Damn it sure takes a lot. I made sure to fix the init now.

Let me make that alternative PG player. I'll subtract the states in the player directly. Getting a good winrate won't be possible for a feedforward player as it won't know what hand does it have.

9:45am.

```
    inl player_pg_alt {name actor learning_rate} cd =
        open Learning
        inl action = RL.action {State Action final=RL.sampling_pg}

        inl input_size = Union.length_dense State
        inl num_actions = Union.length_one_hot Action

        inl net,_ = Tuple.append (Tuple.wrap actor) (RL.Layer.pg {size=num_actions} :: ()) |> init cd input_size
        //inl net,_ = Tuple.append (Tuple.wrap actor) (Feedforward.linear num_actions :: ()) |> init cd input_size

        inl run =
            Union.mutable_function
                (inl {state={state with net} input={input cd}} ->
                    inl input =
                        match state with
                        | {prev_input} -> s.CudaFun.map {map=inl a,b -> a-b} (input,prev_input)
                        | _ -> input
                    inl {action net bck} = action {net input} cd
                    inl bck x =
                        Struct.foldr (inl bck _ -> bck {x with learning_rate}) bck ()
                        Struct.foldr (inl {bck} -> Struct.foldr (inl bck _ -> bck {learning_rate}) bck) net ()
                    {state={net bck prev_input=input}; out=action}
                    )
                {state={net}; input={input=State; cd}}

        inl methods = {basic_methods with
            bet=inl s input -> s.data.run {input cd=s.data.cd}
            showdown=inl s reward ->
                inl l = s.data.run.reset
                inl reward = dyn (to float32 reward)
                List.foldl (inl _ -> function {bck} -> bck {reward} | _ -> ()) () l

                Optimizer.standard learning_rate s.data.cd s.data.net
            game_over=inl s -> ()
            }

        Object
            .member_add methods
            .data_add {name; win=ref 0; net run}
```

Yeah, this should separate out the recurrent nets from the rest. This is still MC training, but let me give it a shot. I want to see what will happen. I'll also try the regular MI nets."