"10am.
```
                    inl input =
                        match state with
                        | {prev_input} ->
                            print_static prev_input.dim
                            cd.CudaFun.map {map=inl a,b -> a-b} (input,prev_input)
                            qwe
                        | _ -> input
                    inl {action net bck} = action {net input} cd
```

Shit, I forgot that the action is not the vector here.

10am. Let me adapt the ac code instead. I do not want to waste too much time on this.

Thunder is coming in. Let me see if I can do an experiment before I am forced to shut down.

10:05am. Oh, the plastic RNN can do it. It is at 697 on the 7th iteration. This is without PRONG in the actor and the critic.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 227 and 773 out of 1000.
The time was 00:00:09.0948644 for: iteration 0
Starting timing for: iteration 1
Winrate is 324 and 676 out of 1000.
The time was 00:00:15.2640695 for: iteration 1
Starting timing for: iteration 2
Winrate is 500 and 500 out of 1000.
The time was 00:00:15.4738740 for: iteration 2
Starting timing for: iteration 3
Winrate is 576 and 424 out of 1000.
The time was 00:00:14.3936064 for: iteration 3
Starting timing for: iteration 4
Winrate is 609 and 391 out of 1000.
The time was 00:00:15.9966000 for: iteration 4
Starting timing for: iteration 5
Winrate is 653 and 347 out of 1000.
The time was 00:00:14.7136173 for: iteration 5
Starting timing for: iteration 6
Winrate is 666 and 334 out of 1000.
The time was 00:00:14.0676806 for: iteration 6
Starting timing for: iteration 7
Winrate is 697 and 303 out of 1000.
The time was 00:00:14.6466612 for: iteration 7
Starting timing for: iteration 8
Winrate is 715 and 285 out of 1000.
The time was 00:00:15.0772665 for: iteration 8
Starting timing for: iteration 9
Winrate is 719 and 281 out of 1000.
The time was 00:00:13.5184832 for: iteration 9
Starting timing for: iteration 10
Winrate is 717 and 283 out of 1000.
The time was 00:00:14.4774382 for: iteration 10
Starting timing for: iteration 11
Winrate is 700 and 300 out of 1000.
The time was 00:00:15.2383081 for: iteration 11
Starting timing for: iteration 12
Winrate is 715 and 285 out of 1000.
The time was 00:00:14.7975664 for: iteration 12
Starting timing for: iteration 13
Winrate is 719 and 281 out of 1000.
The time was 00:00:14.3492459 for: iteration 13
Starting timing for: iteration 14
Winrate is 726 and 274 out of 1000.
The time was 00:00:15.3266523 for: iteration 14
```

Yeah, this is acceptable. Let me try the feedforward net as the baseline."