Today is so fucked up. First the weather stopped me what I wanted to do in the morning and now my bet failed.
I am actually rather depressed over this, but since it has come to this I am starting to come to a conclusion. Maybe the fact that my scheme failed is in fact proof that MC and TD are in fact doing the right thing. It might be proof that internally the states are bound in such a manner that they are separate. This goes both for FF and RNNs. The intuition that starts from tabular representations seems to generalize to distributed ones.

Understanding this in of itself could be an advantage, though this is pretty much the prevailing wisdom. Though it is the prevailing wisdom, it did not at all seem obvious to me though. It seems rather like an enormous stretch that it should generalize from tabular to distributed representations. I felt like a weirdo for being the only one thinking that.

I suppose one could argue that there might be some way to induce the kind of state binding that would make what I had in mind work, but if this did not work, then there is not much point in butting my head against it. I can be a little more confident in standard methods now.

Actually this is the kind of reasoning that could have occurred to me already. I already tried this experiment with standard RNNs and got exactly the same result. Maybe some lightbulbs should have went on back then? I need to readjust my thinking.