In review, apart from not working with standard RNNs I think the biggest clue that the scheme I had in mind would fail would be the learning rates on Mnist. The difference between optimal learning rates on 128 minibatch and online learning is literally 2x. There is no need to average the gradient nearly as much as one would expect. Hence the representation are sparse.
Hence it cannot be the case that the way the standard RL works is by spam.

Rather than worrying about reducing spam I need to turn my thinking around - what I would need is to support sparsity. Good memories would be good for that. This might go hand in hand with chaotic behavior of a NN which would encourage such a thing.

Let me resume the Rudder paper. I've been doing these studies for 3 days now. If it was not for Rudder nor this idea anchoring me, I would have already been long done with the optimized plastic RNN. But I need to do this. Right now I feel like going through the proofs in the paper until I understand them. As papers go, this one is rather imposing so I am going to be studying it for a while. Maybe it will even take me a few days.